+++ yangvalidator/yangvalidator/settings.py
@@ -64,7 +64,7 @@ ALLOWED_HOSTS = [
     'www.yangvalidator.org',
     'yangvalidator.com',
     'yangvalidator.org',
-    '18.224.127.129'
+    '18.184.147.250'
 ]
 
 
@@ -158,11 +158,11 @@ USE_TZ = True
 # Static files (CSS, JavaScript, Images)
 # https://docs.djangoproject.com/en/2.1/howto/static-files/
 
-STATIC_URL = '/yangvalidator/static/'
+STATIC_URL = '/yangvalidator/assets/'
 
-STATICFILES_DIRS = (
-    os.path.join(BASE_DIR, 'static'),
-)
+#STATICFILES_DIRS = (
+#    os.path.join(BASE_DIR, 'static'),
+#)
 
 
 LOGGING_CONFIG = None